# openapi.json
{
  "components": {
    "schemas": {
      "Component": {
        "enum": [
          "interest",
          "ppr",
          "feedback",
          "recency",
          "hide",
          "query"
        ],
        "title": "Component",
        "type": "string"
      },
      "ComponentScore": {
        "properties": {
          "contribution": {
            "title": "Contribution",
            "type": "number"
          },
          "inputs": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Inputs",
            "type": "object"
          },
          "name": {
            "$ref": "#/components/schemas/Component"
          },
          "value": {
            "title": "Value",
            "type": "number"
          },
          "weight": {
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "name",
          "inputs",
          "value",
          "weight",
          "contribution"
        ],
        "title": "ComponentScore",
        "type": "object"
      },
...
      "EventIn": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "impression",
              "click"
            ],
            "title": "Kind",
            "type": "string"
          },
          "position": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "recommendation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Id"
          },
          "surface": {
            "$ref": "#/components/schemas/Surface"
          }
        },
        "required": [
          "kind",
          "document_id",
          "surface"
        ],
        "title": "EventIn",
        "type": "object"
      },
      "EventsIn": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventIn"
            },
            "minItems": 1,
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "EventsIn",
        "type": "object"
      },
      "EventsRecorded": {
        "properties": {
          "recorded": {
            "title": "Recorded",
            "type": "integer"
          }
        },
        "required": [
          "recorded"
        ],
        "title": "EventsRecorded",
        "type": "object"
      },
...
      "FeedItem": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentCard"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/Reason"
            },
            "title": "Reasons",
            "type": "array"
          },
          "recommendation_id": {
            "title": "Recommendation Id",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "slice": {
            "$ref": "#/components/schemas/Slice"
          }
        },
        "required": [
          "recommendation_id",
          "position",
          "slice",
          "score",
          "document",
          "reasons"
        ],
        "title": "FeedItem",
        "type": "object"
      },
      "FeedPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeedItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "FeedPage",
        "type": "object"
      },
      "FeedbackIn": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/FeedbackKind"
          },
          "position": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "reason_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Text"
          },
          "recommendation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Id"
          }
        },
        "required": [
          "document_id",
          "kind"
        ],
        "title": "FeedbackIn",
        "type": "object"
      },
      "FeedbackKind": {
        "enum": [
          "like",
          "hide",
          "block_domain"
        ],
        "title": "FeedbackKind",
        "type": "string"
      },
      "FeedbackOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/FeedbackKind"
          },
          "reason_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Text"
          }
        },
        "required": [
          "id",
          "document_id",
          "kind",
          "reason_text",
          "created_at"
        ],
        "title": "FeedbackOut",
        "type": "object"
      },
      "FeedbackReasonIn": {
        "properties": {
          "reason_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Text"
          }
        },
        "required": [
          "reason_text"
        ],
        "title": "FeedbackReasonIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "Me": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "is_admin": {
            "title": "Is Admin",
            "type": "boolean"
          },
          "survey_completed": {
            "title": "Survey Completed",
            "type": "boolean"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "email",
          "timezone",
          "survey_completed",
          "is_admin"
        ],
        "title": "Me",
        "type": "object"
      },
...
      "Reason": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ReasonKind"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "text"
        ],
        "title": "Reason",
        "type": "object"
      },
      "ReasonKind": {
        "enum": [
          "exploration",
          "sources",
          "interest",
          "liked",
          "recency",
          "search"
        ],
        "title": "ReasonKind",
        "type": "string"
      },
      "Slice": {
        "enum": [
          "main",
          "adjacent_semantic",
          "adjacent_graph"
        ],
        "title": "Slice",
        "type": "string"
      },
...
      "Surface": {
        "enum": [
          "feed",
          "search"
        ],
        "title": "Surface",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Why": {
        "properties": {
          "components": {
            "items": {
              "$ref": "#/components/schemas/ComponentScore"
            },
            "title": "Components",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document": {
            "$ref": "#/components/schemas/DocumentCard"
          },
          "evidence": {
            "$ref": "#/components/schemas/Evidence"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/Reason"
            },
            "title": "Reasons",
            "type": "array"
          },
          "recommendation_id": {
            "title": "Recommendation Id",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "slice": {
            "$ref": "#/components/schemas/Slice"
          },
          "surface": {
            "$ref": "#/components/schemas/Surface"
          }
        },
        "required": [
          "recommendation_id",
          "surface",
          "query",
          "slice",
          "score",
          "created_at",
          "document",
          "reasons",
          "components",
          "evidence"
        ],
        "title": "Why",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Discovery Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/session": {
      "post": {
        "description": "Trade a one-time login token for a session cookie.",
        "operationId": "sign_in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign In",
        "tags": [
          "auth"
        ]
      }
    },
    "/events": {
      "post": {
        "description": "Record a batch of impressions and clicks.",
        "operationId": "record_events",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsRecorded"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Events",
        "tags": [
          "feedback"
        ]
      }
    },
    "/feed": {
      "get": {
        "description": "The next page of the user's feed. Without a cursor, a new feed session starts.",
        "operationId": "feed",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feed",
        "tags": [
          "feed"
        ]
      }
    },
    "/feedback": {
      "post": {
        "description": "Like, hide, or block the document's domain. Hidden documents and blocked domains leave\nthe feed at once; likes and hides reshape the profile vectors right away too.",
        "operationId": "give_feedback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Give Feedback",
        "tags": [
          "feedback"
        ]
      }
    },
    "/feedback/{feedback_id}": {
      "patch": {
        "description": "Add (or clear) the optional reason after the fact: the \"What did you like about it?\"\nprompt comes after the like.",
        "operationId": "set_feedback_reason",
        "parameters": [
          {
            "in": "path",
            "name": "feedback_id",
            "required": true,
            "schema": {
              "title": "Feedback Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackReasonIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Feedback Reason",
        "tags": [
          "feedback"
        ]
      }
    },
...
    "/recommendations/{recommendation_id}/why": {
      "get": {
        "description": "The full score breakdown of a served item (PLAN.md \u00a76.7). Logs a `why_open` event.",
        "operationId": "why",
        "parameters": [
          {
            "in": "path",
            "name": "recommendation_id",
            "required": true,
            "schema": {
              "title": "Recommendation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Why"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Why",
        "tags": [
          "feed"
        ]
      }
    },
    "/search": {
      "get": {
        "description": "Documents matching the query, ranked with the user's own signals on top.",
        "operationId": "search",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "feed"
        ]
      }
    },
...
  }
}